+++ omp_tree_barrier.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdint.h>
+#include <assert.h>
 #include "omp_tree_barrier.h"
 
 int construct_tree_aux(int start_id, int end_id, omp_tree_barrier_node_t* parent, omp_tree_barrier_node_t* nodes,
@@ -39,7 +40,7 @@ void omp_tree_barrier_init(omp_tree_barrier_t *barrier, int num_threads) {
   posix_memalign((void**)&barrier->nodes, 64, sizeof(omp_tree_barrier_node_t) * barrier->N);
 
 #ifdef CACHE_PADDING
-  assert((barrier->nodes % 64) == 0)
+  assert(((long)barrier->nodes % 64) == 0)
   assert(sizeof(omp_tree_barrier_node_t) == 64);
 #endif
   if (barrier->nodes == NULL) {
@@ -49,7 +50,7 @@ void omp_tree_barrier_init(omp_tree_barrier_t *barrier, int num_threads) {
 
   posix_memalign((void**)&barrier->threads, 64, sizeof(tree_barrier_thread_data_t) * barrier->N);
 #ifdef CACHE_PADDING
-  assert((barrier->threads % 64) == 0)
+  assert(((long)barrier->threads % 64) == 0)
   assert(sizeof(tree_barrier_thread_data_t) == 64);
 #endif
   if (barrier->threads == NULL) {
